requirements.txt
camelot-py==0.7.2
chardet==3.0.4
Click==7.0
decorator==4.4.0
Elixir==0.7.1
et-xmlfile==1.0.1
jdcal==1.4.1
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.16.4
opencv-python==4.1.0.25
openpyxl==2.6.2
pandas==0.24.2
pbr==5.2.1
pdf-to-markdown==0.1.0
pdfminer.six==20181108
pkg-resources==0.0.0
pycryptodome==3.8.1
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2019.1
selenium==3.141.0
six==1.12.0
sortedcontainers==2.1.0
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.12.0
sqlparse==0.3.0
Tempita==0.5.2
urllib3==1.25.3
xlrd==0.7.1
xlwt==0.7.2
